There are other common use cases for PROMPT_COMMAND than setting PS1.… (#22)
* There are other common use cases for PROMPT_COMMAND than setting PS1. For example:

- [`direnv`](https://direnv.net/) loads rc files upon a change of directory
- `/etc/profile.d/vte.sh` also sets this variable on some operating systems, e.g. Fedora Linux, to make it possible to open new virtual terminal tabs in the same directory as the current one.

On my system, before loading powerbash:

```bash
echo ${PROMPT_COMMAND}
```

gives

```
_direnv_hook;__vte_prompt_command
```

The modified code keeps existing settings. They are extended instead of overwritten when loading powerbash. As of Bash 5.1, PROMPT_COMMAND can be an array, which is also supported in this change.

Note that the `export` from `export PROMPT_COMMAND` was removed. When PROMPT_COMMAND is an array (newer bash versions) it cannot be exported, so one should not rely on it. This also means, at least on most Linux systems, that powerbash should be enabled in `.bashrc`, not `.bash_profile`. One cannot assume that the PROMPT_COMMAND array persists in subprocess, so it needs to be reinitialized every time bash starts. (The installer script puts it in .bash_profile, which did not work for me.)

* Append instead of prepend